pyproject.toml
[tool.poetry]
name = "asr-task-automation-ec"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
mirascope = "^0.18.1"
requests = "^2.32.3"
streamlit = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
